[G08] : [MOVE] Move Group 08 test.js API file into graderCreate Folder
